+++ compiled/spec/unit/event-listener.decorator.spec.js
@@ -25,10 +25,13 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var src_1 = require("../../src");
 describe('Event Listener Decorator', function () {
     it('should bind and unbind to events', function () {
+        var HUB_EVENTS = {
+            CNT_UP: 'cnt-up'
+        };
         var Hub = (function (_super) {
             __extends(Hub, _super);
             function Hub() {
-                var _this = _super.call(this) || this;
+                var _this = _super.call(this, HUB_EVENTS) || this;
                 _this.cnt = 0;
                 _this.init();
                 return _this;
@@ -40,7 +43,7 @@ describe('Event Listener Decorator', function () {
             };
             __decorate([
                 src_1.EventListener({
-                    eventName: 'cnt-up',
+                    eventName: HUB_EVENTS.CNT_UP,
                     initFn: 'init',
                     destroyFn: 'destroy'
                 }),
@@ -51,17 +54,20 @@ describe('Event Listener Decorator', function () {
             return Hub;
         }(src_1.EventManager));
         var hub = new Hub();
-        hub.emit('cnt-up');
+        hub.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(1);
         hub.destroy();
-        hub.emit('cnt-up');
+        hub.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(1);
     });
     it('should use default angular init and destroy functions', function () {
+        var HUB_EVENTS = {
+            CNT_UP: 'cnt-up'
+        };
         var Hub = (function (_super) {
             __extends(Hub, _super);
             function Hub() {
-                var _this = _super !== null && _super.apply(this, arguments) || this;
+                var _this = _super.call(this, HUB_EVENTS) || this;
                 _this.cnt = 0;
                 return _this;
             }
@@ -71,7 +77,7 @@ describe('Event Listener Decorator', function () {
                 this.cnt++;
             };
             __decorate([
-                src_1.EventListener('cnt-up'),
+                src_1.EventListener(HUB_EVENTS.CNT_UP),
                 __metadata("design:type", Function),
                 __metadata("design:paramtypes", []),
                 __metadata("design:returntype", void 0)
@@ -80,17 +86,20 @@ describe('Event Listener Decorator', function () {
         }(src_1.EventManager));
         var hub = new Hub();
         hub['ngOnInit']();
-        hub.emit('cnt-up');
+        hub.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(2);
         hub['ngOnDestroy']();
-        hub.emit('cnt-up');
+        hub.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(2);
     });
     it('should use property class', function () {
+        var HUB_EVENTS = {
+            CNT_UP: 'cnt-up'
+        };
         var EventHub = (function (_super) {
             __extends(EventHub, _super);
             function EventHub() {
-                return _super !== null && _super.apply(this, arguments) || this;
+                return _super.call(this, HUB_EVENTS) || this;
             }
             return EventHub;
         }(src_1.EventManager));
@@ -106,11 +115,11 @@ describe('Event Listener Decorator', function () {
                 this.cnt++;
             };
             MyTest.prototype.count = function () {
-                this.eventHub.emit('cnt-up');
+                this.eventHub.emit(HUB_EVENTS.CNT_UP);
             };
             __decorate([
                 src_1.EventListener({
-                    eventName: 'cnt-up',
+                    eventName: HUB_EVENTS.CNT_UP,
                     eventClass: EventHub.name,
                     initFn: 'init',
                     destroyFn: 'destroy'
@@ -133,10 +142,13 @@ describe('Event Listener Decorator', function () {
     it('should throw error as class name cant be found', function () {
         var err = null;
         try {
+            var HUB_EVENTS_1 = {
+                CNT_UP: 'cnt-up'
+            };
             var EventHub = (function (_super) {
                 __extends(EventHub, _super);
                 function EventHub() {
-                    return _super !== null && _super.apply(this, arguments) || this;
+                    return _super.call(this, HUB_EVENTS_1) || this;
                 }
                 return EventHub;
             }(src_1.EventManager));
@@ -152,11 +164,11 @@ describe('Event Listener Decorator', function () {
                     this.cnt++;
                 };
                 MyTest.prototype.count = function () {
-                    this.eventHub.emit('cnt-up');
+                    this.eventHub.emit(HUB_EVENTS_1.CNT_UP);
                 };
                 __decorate([
                     src_1.EventListener({
-                        eventName: 'cnt-up',
+                        eventName: HUB_EVENTS_1.CNT_UP,
                         eventClass: 'fake-name',
                         initFn: 'init',
                         destroyFn: 'destroy'
@@ -207,15 +219,18 @@ describe('Event Listener Decorator', function () {
             err = e;
         }
         expect(err.message)
-            .toContain('@EventListener: Class Hub must extend EventManager');
+            .toContain('@EventListener: No EventManager class');
     });
     it('should throw error as decorator is not on a method', function () {
         var err = null;
         try {
+            var HUB_EVENTS_2 = {
+                CNT_UP: 'cnt-up'
+            };
             var Hub = (function (_super) {
                 __extends(Hub, _super);
                 function Hub() {
-                    var _this = _super.call(this) || this;
+                    var _this = _super.call(this, HUB_EVENTS_2) || this;
                     _this.cnt = 0;
                     _this.cntListener = true;
                     _this.init();
@@ -225,7 +240,7 @@ describe('Event Listener Decorator', function () {
                 Hub.prototype.destroy = function () { };
                 __decorate([
                     src_1.EventListener({
-                        eventName: 'cnt-up',
+                        eventName: HUB_EVENTS_2.CNT_UP,
                         initFn: 'init',
                         destroyFn: 'destroy'
                     }),
@@ -244,10 +259,13 @@ describe('Event Listener Decorator', function () {
     it('should throw error as args are not set', function () {
         var err = null;
         try {
+            var HUB_EVENTS_3 = {
+                CNT_UP: 'cnt-up'
+            };
             var Hub = (function (_super) {
                 __extends(Hub, _super);
                 function Hub() {
-                    var _this = _super.call(this) || this;
+                    var _this = _super.call(this, HUB_EVENTS_3) || this;
                     _this.cnt = 0;
                     _this.init();
                     return _this;
@@ -273,10 +291,18 @@ describe('Event Listener Decorator', function () {
             + 'type string or IEventListenerArgs');
     });
     it('should bind and unbind to events on the events property', function () {
+        var HUB_EVENTS = {
+            CNT_UP: 'cnt-up'
+        };
+        var USER_EVENTS = {
+            ADDED: 'user-added'
+        };
         var Hub = (function () {
             function Hub() {
-                this.events = new src_1.EventManager();
+                this.events = new src_1.EventManager(HUB_EVENTS);
+                this.userEvents = new src_1.EventManager(USER_EVENTS);
                 this.cnt = 0;
+                this.usersAdded = 0;
                 this.init();
             }
             Hub.prototype.init = function () { };
@@ -284,9 +310,12 @@ describe('Event Listener Decorator', function () {
             Hub.prototype.cntListener = function () {
                 this.cnt++;
             };
+            Hub.prototype.usersAddListener = function () {
+                this.usersAdded++;
+            };
             __decorate([
                 src_1.EventListener({
-                    eventName: 'cnt-up',
+                    eventName: HUB_EVENTS.CNT_UP,
                     initFn: 'init',
                     destroyFn: 'destroy'
                 }),
@@ -294,16 +323,31 @@ describe('Event Listener Decorator', function () {
                 __metadata("design:paramtypes", []),
                 __metadata("design:returntype", void 0)
             ], Hub.prototype, "cntListener", null);
+            __decorate([
+                src_1.EventListener({
+                    eventName: USER_EVENTS.ADDED,
+                    initFn: 'init',
+                    destroyFn: 'destroy'
+                }),
+                __metadata("design:type", Function),
+                __metadata("design:paramtypes", []),
+                __metadata("design:returntype", void 0)
+            ], Hub.prototype, "usersAddListener", null);
             return Hub;
         }());
         var hub = new Hub();
-        hub.events.emit('cnt-up');
+        hub.events.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(1);
+        hub.userEvents.emit(USER_EVENTS.ADDED);
+        expect(hub.usersAdded).toEqual(1);
         hub.destroy();
-        hub.events.emit('cnt-up');
+        hub.events.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(1);
     });
     it('should bind and unbind to events on random name property', function () {
+        var HUB_EVENTS = {
+            CNT_UP: 'cnt-up'
+        };
         var EventFake = (function () {
             function EventFake() {
             }
@@ -314,7 +358,7 @@ describe('Event Listener Decorator', function () {
         var Hub = (function () {
             function Hub() {
                 this.someProp = new EventFake();
-                this.randomName = new src_1.EventManager();
+                this.randomName = new src_1.EventManager(HUB_EVENTS);
                 this.cnt = 0;
                 this.init();
             }
@@ -325,7 +369,7 @@ describe('Event Listener Decorator', function () {
             };
             __decorate([
                 src_1.EventListener({
-                    eventName: 'cnt-up',
+                    eventName: HUB_EVENTS.CNT_UP,
                     initFn: 'init',
                     destroyFn: 'destroy'
                 }),
@@ -336,10 +380,10 @@ describe('Event Listener Decorator', function () {
             return Hub;
         }());
         var hub = new Hub();
-        hub.randomName.emit('cnt-up');
+        hub.randomName.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(1);
         hub.destroy();
-        hub.randomName.emit('cnt-up');
+        hub.randomName.emit(HUB_EVENTS.CNT_UP);
         expect(hub.cnt).toEqual(1);
     });
     it('should throw error as the events property is not of type EventManager', function () {
@@ -373,11 +417,14 @@ describe('Event Listener Decorator', function () {
         catch (e) {
             err = e;
         }
-        expect(err.message).toContain('@EventListener: Class');
+        expect(err.message).toContain('@EventListener: No EventManager class');
     });
     it('should throw error as ngOnDestroy does not exist', function () {
         var err = null;
         try {
+            var HUB_EVENTS_4 = {
+                CNT_UP: 'cnt-up'
+            };
             var Hub = (function (_super) {
                 __extends(Hub, _super);
                 function Hub() {
@@ -390,14 +437,14 @@ describe('Event Listener Decorator', function () {
                     this.cnt++;
                 };
                 __decorate([
-                    src_1.EventListener('cnt-up'),
+                    src_1.EventListener(HUB_EVENTS_4.CNT_UP),
                     __metadata("design:type", Function),
                     __metadata("design:paramtypes", []),
                     __metadata("design:returntype", void 0)
                 ], Hub.prototype, "cntListener", null);
                 return Hub;
             }(src_1.EventManager));
-            var hub = new Hub();
+            var hub = new Hub(HUB_EVENTS_4);
         }
         catch (e) {
             err = e;
@@ -407,6 +454,9 @@ describe('Event Listener Decorator', function () {
     it('should throw error as ngOnInit does not exist', function () {
         var err = null;
         try {
+            var HUB_EVENTS_5 = {
+                CNT_UP: 'cnt-up'
+            };
             var Hub = (function (_super) {
                 __extends(Hub, _super);
                 function Hub() {
@@ -419,14 +469,14 @@ describe('Event Listener Decorator', function () {
                     this.cnt++;
                 };
                 __decorate([
-                    src_1.EventListener('cnt-up'),
+                    src_1.EventListener(HUB_EVENTS_5.CNT_UP),
                     __metadata("design:type", Function),
                     __metadata("design:paramtypes", []),
                     __metadata("design:returntype", void 0)
                 ], Hub.prototype, "cntListener", null);
                 return Hub;
             }(src_1.EventManager));
-            var hub = new Hub();
+            var hub = new Hub(HUB_EVENTS_5);
         }
         catch (e) {
             err = e;
@@ -464,6 +514,118 @@ describe('Event Listener Decorator', function () {
             err = e;
         }
         expect(err.message).toContain('both initFn and destroyFn');
+    });
+    it('should work with multiple classes of same type', function () {
+        var EventFake = (function () {
+            function EventFake() {
+            }
+            EventFake.prototype.on = function () { };
+            return EventFake;
+        }());
+        ;
+        var ONE_EVENTS = {
+            SOMETHING_ADDED: 's1-something-added'
+        };
+        var ServiceOne = (function () {
+            function ServiceOne(e) {
+                this.events = new src_1.EventManager(e);
+            }
+            return ServiceOne;
+        }());
+        var TWO_EVENTS = {
+            SOMETHING_ADDED: 's2-something-added'
+        };
+        var ServiceTwo = (function () {
+            function ServiceTwo(e) {
+                this.events = new src_1.EventManager(e);
+            }
+            return ServiceTwo;
+        }());
+        var Hub = (function () {
+            function Hub() {
+                this.serviceOne = new ServiceOne(ONE_EVENTS);
+                this.fakeEventClass = new EventFake();
+                this.serviceTwoButDifferent = new ServiceTwo(ONE_EVENTS);
+                this.serviceTwo = new ServiceTwo(TWO_EVENTS);
+                this.cnt = 0;
+            }
+            Hub.prototype.ngOnInit = function () { };
+            Hub.prototype.ngOnDestroy = function () { };
+            Hub.prototype.cntListener = function () {
+                this.cnt++;
+            };
+            __decorate([
+                src_1.EventListener(TWO_EVENTS.SOMETHING_ADDED, ServiceTwo),
+                __metadata("design:type", Function),
+                __metadata("design:paramtypes", []),
+                __metadata("design:returntype", void 0)
+            ], Hub.prototype, "cntListener", null);
+            return Hub;
+        }());
+        var hub = new Hub();
+        hub.ngOnInit();
+        hub.serviceTwo.events.emit(TWO_EVENTS.SOMETHING_ADDED);
+        expect(hub.cnt).toEqual(1);
+        hub.serviceOne.events.emit(ONE_EVENTS.SOMETHING_ADDED);
+        expect(hub.cnt).toEqual(1);
+        hub.ngOnDestroy();
+    });
+    it('should throw error as two classes use same event name', function () {
+        var EventFake = (function () {
+            function EventFake() {
+            }
+            EventFake.prototype.on = function () { };
+            return EventFake;
+        }());
+        ;
+        var ONE_EVENTS = {
+            SOMETHING_ADDED: 'something-added'
+        };
+        var ServiceOne = (function () {
+            function ServiceOne(e) {
+                this.events = new src_1.EventManager(e);
+            }
+            return ServiceOne;
+        }());
+        var TWO_EVENTS = {
+            SOMETHING_ADDED: 'something-added'
+        };
+        var ServiceTwo = (function () {
+            function ServiceTwo(e) {
+                this.events = new src_1.EventManager(e);
+            }
+            return ServiceTwo;
+        }());
+        var Hub = (function () {
+            function Hub() {
+                this.serviceOne = new ServiceOne(ONE_EVENTS);
+                this.fakeEventClass = new EventFake();
+                this.serviceTwoButDifferent = new ServiceTwo(ONE_EVENTS);
+                this.serviceTwo = new ServiceTwo(TWO_EVENTS);
+                this.cnt = 0;
+            }
+            Hub.prototype.ngOnInit = function () { };
+            Hub.prototype.ngOnDestroy = function () { };
+            Hub.prototype.cntListener = function () {
+                this.cnt++;
+            };
+            __decorate([
+                src_1.EventListener(TWO_EVENTS.SOMETHING_ADDED, ServiceTwo),
+                __metadata("design:type", Function),
+                __metadata("design:paramtypes", []),
+                __metadata("design:returntype", void 0)
+            ], Hub.prototype, "cntListener", null);
+            return Hub;
+        }());
+        var err;
+        try {
+            var hub = new Hub();
+            hub.ngOnInit();
+        }
+        catch (e) {
+            err = e;
+        }
+        expect(err.message).toContain('Found 2 or more classes');
     });
 });
 //# sourceMappingURL=event-listener.decorator.spec.js.map