+++ compiled/src/event-manager.d.ts
@@ -3,8 +3,13 @@ export declare type INextFn = () => {
     completed: (cb: () => void) => void;
 };
 export declare class EventManager {
+    private emittedEvents;
     private _events;
     private events;
+    constructor(emittedEvents: {
+        [key: string]: string;
+    });
+    eventExists(name: string): boolean;
     emit(eventName: string, data?: any): void;
     on(eventName: string, fn: ICallbackFn, scope?: any): void;
     once(eventName: string, fn: ICallbackFn, scope?: any): void;